+++ src/pyrite/rendering/__init__.py
@@ -3,12 +3,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from .bounds_service import BoundsService  # noqa: F401
-from .camera_renderer import CameraRenderer  # noqa: F401
 from .ortho_projection import OrthoProjection  # noqa: F401
 from .rect_bounds import RectBounds  # noqa: F401
 from .render_texture import RenderTexture, RenderTextureComponent  # noqa: F401
-from .sprite_renderer import SpriteRenderer  # noqa: F401
-from .view_plane import ViewPlane  # noqa: F401
 from .viewport import Viewport  # noqa: F401
 
 if TYPE_CHECKING:
